requirements.txt
redis>=7.1.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
orjson>=3.11.0
structlog>=25.5.0
opentelemetry-api>=1.38.0